refactor: move class out of domain
